requirements.txt
openai>=1.30.0
faiss-cpu>=1.7.4
streamlit>=1.35.0
numpy>=1.26.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0